requirements.txt
# Core LangChain and LangGraph
langchain>=0.1.0
langchain-core>=0.1.0
langchain-upstage>=0.1.0
langgraph>=0.1.0

# Free Embedding Models (no OpenAI)
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0

# Vector Store
faiss-cpu>=1.7.4

# Web Search
tavily-python>=0.3.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
alembic>=1.12.0

# Streamlit and Visualization
streamlit>=1.28.0
streamlit-chat>=0.1.1
plotly>=5.17.0
altair>=5.0.0
streamlit-option-menu>=0.3.6

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
tqdm>=4.65.0

# Development
jupyter>=1.0.0
pytest>=7.4.0
black>=23.0.0

matplotlib
